feat: redisConfig 위치 이동 및 수정
